backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
tavily
bibtexparser
python-dotenv
feedparser
requests